DAAC-441: Prepare images and labels in Tensorflow dataset
The Tensorflow API can be used to batch these data and prefetch batches during training. This reduces memory pressure during training by reducing the number of batches that need to be loaded into memory at once.

This menas that the training pipeline can be run locally on a development machine without being oom-killed. Additionally the training pipeline is speed up significantly as we are not swapping heavily.